Use halt instead of poweroff when stopping machines
This helps to drain the caps when the power is cut.
This only halts physical machines. Virtual machines still use poweroff.

+++ src/machine.php
@@ -513,8 +513,10 @@ class Machine {
 		$status = $this->get_status();
 
 		if ($status == "online") {
-			out("Executing poweroff on machine");
-			$this->ssh_cmd("poweroff");
+			if ($this->is_vm())
+				$this->ssh_cmd("poweroff");
+			else
+				$this->ssh_cmd("halt");
 		}
 
 		if ($status == "unreachable") {
